refactor(repository): chatMessage repository 통합
- fix: cursorId도 응답에 포함되도록 수정(반환 값에 nextId가 포함되므로)
