+++ src/main/java/com/pnu/momeet/domain/chatting/repository/ChatMessageRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.UUID;
 
-public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
+public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long>, ChatMessageDslRepository {
 
     @Modifying
     @Query("DELETE FROM ChatMessage cm WHERE cm.meetup.id = :meetupId")
